requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.1
filelock==3.15.4
fsspec==2024.6.1
grpcio==1.65.5
grpcio-tools==1.65.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
protobuf==5.27.3
pydantic==2.8.2
pydantic_core==2.20.1
python-multipart==0.0.9
PyYAML==6.0.2
qdrant-client==1.11.0
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sniffio==1.3.1
starlette==0.38.2
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
